move things around to prepare for error handling
